+++ Polyfills/Canvas/Source/Context.h
@@ -3,6 +3,7 @@
 #include <Babylon/Polyfills/Canvas.h>
 #include <Babylon/JsRuntimeScheduler.h>
 #include <GraphicsImpl.h>
+#include "Image.h"
 
 struct NVGcontext;
 
@@ -72,8 +73,7 @@ namespace Babylon::Polyfills::Internal
         void Dispose(const Napi::CallbackInfo&);
         void Dispose();
         void SetDirty();
-        void BeginFrame();
-        void EndFrame();
+        void DeferredFlushFrame();
 
         NativeCanvas* m_canvas;
         NVGcontext* m_nvg;
@@ -90,6 +90,8 @@ namespace Babylon::Polyfills::Internal
         bool m_dirty{};
         std::shared_ptr<arcana::cancellation_source> m_cancellationSource{};
         JsRuntimeScheduler m_runtimeScheduler;
+
+        std::unordered_map<const NativeCanvasImage*, int> m_nvgImageIndices;
 
         void FlushGraphicResources() override;
